+++ src/DependencyInjection/RichIdCookiesRegulationExtension.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace RichId\CookiesRegulationBundle\DependencyInjection;
 
 use RichCongress\BundleToolbox\Configuration\AbstractExtension;
-use RichId\CookiesRegulationBundle\DependencyInjection\PrependTrait\DoctrineMigrationsPrependTrait;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 
 /**
@@ -16,10 +14,8 @@ use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
  */
-class RichIdCookiesRegulationExtension extends AbstractExtension implements PrependExtensionInterface
+class RichIdCookiesRegulationExtension extends AbstractExtension
 {
-    use DoctrineMigrationsPrependTrait;
-
     /** @param array<string, mixed> $configs */
     public function load(array $configs, ContainerBuilder $container): void
     {
@@ -31,10 +27,5 @@ class RichIdCookiesRegulationExtension extends AbstractExtension implements Prep
 
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources'));
         $loader->load('services.xml');
-    }
-
-    public function prepend(ContainerBuilder $container): void
-    {
-        $this->prependDoctrineMigrations($container);
     }
 }