+++ src/RichIdCookiesRegulationBundle.php
@@ -4,32 +4,12 @@ declare(strict_types=1);
 
 namespace RichId\CookiesRegulationBundle;
 
-use Doctrine\Bundle\DoctrineBundle\DependencyInjection\Compiler\DoctrineOrmMappingsPass;
 use RichCongress\BundleToolbox\Configuration\AbstractBundle;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 class RichIdCookiesRegulationBundle extends AbstractBundle
 {
-    public const COMPILER_PASSES = [];
-
-    public function build(ContainerBuilder $container): void
-    {
-        parent::build($container);
-
-        $this->addDoctrineOrmMappingsPass($container);
-    }
-
-    private function addDoctrineOrmMappingsPass(ContainerBuilder $container): void
-    {
-        if (!\class_exists(DoctrineOrmMappingsPass::class)) {
-            return;
-        }
-
-        $container->addCompilerPass(
-            DoctrineOrmMappingsPass::createAnnotationMappingDriver(
-                ['RichId\CookiesRegulationBundle\Entity'],
-                [__DIR__ . '/Entity']
-            )
-        );
-    }
+    /** @var array<string, string> */
+    protected static $doctrineAnnotationMapping = [
+        'RichId\CookiesRegulationBundle\Entity' => __DIR__ . '/Entity',
+    ];
 }