+++ tests/Resources/Kernel/config/bundles.php
@@ -3,5 +3,6 @@
 declare(strict_types=1);
 
 return [
+    Symfony\Bundle\MonologBundle\MonologBundle::class                   => ['all' => true],
     RichId\CookiesRegulationBundle\RichIdCookiesRegulationBundle::class => ['all' => true],
 ];