+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## Version 0.1.0
+
+- Initial commit
+- Add enable option in the service configuration
+- Add mandatory option in the service configuration